+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.1.1 (2018-04-02)
+
+* Fix: `getItemAvailability` only read first passed in external/internal ID
+* Feature: Allow overriding global NS preferences through SOAP headers
+
 ## 0.1.0 (2018-03-29)
 
 * Initial version. Support for `getList` and `getItemAvailability`